Update UI: extended MyCursorAdapter Code: solved minor bags

+++ src/com/example/silencer/TaskPaneEdit.java
@@ -46,7 +46,7 @@ public class TaskPaneEdit extends Activity implements TimeDialogListener{
 	
 	DBHelper dbHelper;
 	DialogTime dialogtime;
-	long id;
+	int id;
 	SQLiteDatabase db;
 	
 	Time timeStart;	
@@ -78,7 +78,7 @@ public class TaskPaneEdit extends Activity implements TimeDialogListener{
 		Intent intent = getIntent();
 	    
 		//row number
-		id = intent.getLongExtra("id", 0);
+		id = intent.getIntExtra("id", 0);
         Log.d(LOG_TAG, "row selected, ID = " + id);
 	    dialogtime = new DialogTime();
